+++ CHANGELOG.md
@@ -2,10 +2,19 @@
 
 ## Current Version
 
+v3.0.x
+
+- Remove use of static classes for TokenEstimator
+- Added `Logger` to `ModelTokenizer`
+
+## Previous Versions
+
+v2.0.x
+
+- Compatibility with v1.0.0
+
 v1.0.x
 
 - Initial release
-
-## Previous Versions
 
 Notes from previous versions will be listed here.